New version of find_rv.py for outlier rejection
This new version allows you to set a window for outlier rejection AND force 1000 iterations to be created in it. On the plus side, this means all RV histograms will be fully sampled. On the minus side, this could run effectively forever if the window limits are set to an empty portion of pixel-shift space.